flask_backend/requirements.txt
flask
flask-cors
gunicorn
tensorflow
tensorflow-text
numpy
pandas
tf-models-official
scikit-learn
bert-tensorflow
nltk
tensorflow-hub
firebase
firebase-admin
matplotlib
oauth2client
google
google-api-python-client
openapi
